+++ src/connector.ts
@@ -12,7 +12,7 @@ import { CljsTypeConfig, ReplConnectSequence, getDefaultCljsType, CljsTypes, ask
 import { disabledPrettyPrinter } from './printer';
 import { keywordize } from './util/string';
 import { REQUESTS, initializeDebugger } from './debugger/calva-debug';
-import * as resultsOutput from './result-output'
+import * as outputWindow from './result-output'
 import evaluate from './evaluate';
 
 async function createAndConnectReplWindow(session: NReplSession, mode: "clj" | "cljs"): Promise<void> {
@@ -43,14 +43,14 @@ async function connectToHost(hostname, port, connectSequence: ReplConnectSequenc
     util.setConnectingState(true);
     status.update();
     try {
-        resultsOutput.appendToResultsDoc("; Hooking up nREPL sessions...");
+        outputWindow.appendToResultsDoc("; Hooking up nREPL sessions...");
         // Create an nREPL client. waiting for the connection to be established.
         nClient = await NReplClient.create({ host: hostname, port: +port })
         nClient.addOnCloseHandler(c => {
             util.setConnectedState(false);
             util.setConnectingState(false);
             if (!c["silent"]) // we didn't deliberately close this session, mention this fact.
-                resultsOutput.appendToResultsDoc("; nREPL Connection was closed");
+                outputWindow.appendToResultsDoc("; nREPL Connection was closed");
             status.update();
         })
         cljSession = nClient.session;
@@ -61,31 +61,31 @@ async function connectToHost(hostname, port, connectSequence: ReplConnectSequenc
         state.cursor.set('clj', cljSession);
         state.cursor.set('cljc', cljSession);
         status.update();
-        const outputDocument = await resultsOutput.openResultsDoc();
-        resultsOutput.appendToResultsDoc(`; Connected session: clj\n${resultsOutput.CLJ_CONNECT_GREETINGS}`);
-        resultsOutput.setSession(cljSession, nClient.ns);
+        outputWindow.appendToResultsDoc(`; Connected session: clj\n${outputWindow.CLJ_CONNECT_GREETINGS}`);
+        outputWindow.setSession(cljSession, nClient.ns);
         util.updateREPLSessionType();
 
         // Initialize debugger
         await initializeDebugger(cljSession);
-        resultsOutput.appendToResultsDoc('; Debugger initialized');
+        outputWindow.appendToResultsDoc('; Debugger initialized');
 
         await createAndConnectReplWindow(cljSession, "clj");
 
         if (connectSequence.afterCLJReplJackInCode) {
+            const outputDocument = await outputWindow.openResultsDoc();
             const evalPos = outputDocument.positionAt(outputDocument.getText().length);
-            await resultsOutput.appendToResultsDoc(`; Evaluating 'afterCLJReplJackInCode'\n${connectSequence.afterCLJReplJackInCode}`);
+            await outputWindow.appendToResultsDoc(`; Evaluating 'afterCLJReplJackInCode'\n${connectSequence.afterCLJReplJackInCode}`);
             try {
                 await evaluate.evaluateCode(connectSequence.afterCLJReplJackInCode, {
                     filePath: outputDocument.fileName,
-                    session: resultsOutput.getSession(),
-                    ns: resultsOutput.getNs(),
+                    session: outputWindow.getSession(),
+                    ns: outputWindow.getNs(),
                     line: evalPos.line,
                     column: evalPos.character
                 });
             }
             catch (e) {
-                resultsOutput.appendToResultsDoc("; Evaluation of afterCLJReplJackInCode failed.")
+                outputWindow.appendToResultsDoc("; Evaluation of afterCLJReplJackInCode failed.")
             }
         }
 
@@ -104,13 +104,13 @@ async function connectToHost(hostname, port, connectSequence: ReplConnectSequenc
                 await setUpCljsRepl(cljsSession, cljsBuild);
             }
         } catch (e) {
-            resultsOutput.appendToResultsDoc("; Error while connecting cljs REPL: " + e);
+            outputWindow.appendToResultsDoc("; Error while connecting cljs REPL: " + e);
         }
         status.update();
     } catch (e) {
         util.setConnectingState(false);
         util.setConnectedState(false);
-        resultsOutput.appendToResultsDoc("; Failed connecting.");
+        outputWindow.appendToResultsDoc("; Failed connecting.");
         state.analytics().logEvent("REPL", "FailedConnectingCLJ").send();
         return false;
     }
@@ -121,8 +121,8 @@ async function connectToHost(hostname, port, connectSequence: ReplConnectSequenc
 async function setUpCljsRepl(session, build) {
     state.cursor.set("cljs", session);
     status.update();
-    resultsOutput.appendToResultsDoc(`; Connected session: cljs${(build ? ", repl: " + build : "")}\n${resultsOutput.CLJS_CONNECT_GREETINGS}`);
-    resultsOutput.setSession(session, 'cljs.user');
+    outputWindow.appendToResultsDoc(`; Connected session: cljs${(build ? ", repl: " + build : "")}\n${outputWindow.CLJS_CONNECT_GREETINGS}`);
+    outputWindow.setSession(session, 'cljs.user');
     util.updateREPLSessionType();
     createAndConnectReplWindow(session, "cljs");
 }
@@ -132,8 +132,8 @@ function getFigwheelMainBuilds() {
     let builds = res.filter(x => x.match(/\.cljs\.edn/)).map(x => x.replace(/\.cljs\.edn$/, ""));
     if (builds.length == 0) {
         vscode.window.showErrorMessage("There are no figwheel build files (.cljs.edn) in the project directory.");
-        resultsOutput.appendToResultsDoc("; There are no figwheel build files (.cljs.edn) in the project directory.");
-        resultsOutput.appendToResultsDoc("; Connection to Figwheel Main aborted.");
+        outputWindow.appendToResultsDoc("; There are no figwheel build files (.cljs.edn) in the project directory.");
+        outputWindow.appendToResultsDoc("; Connection to Figwheel Main aborted.");
         throw "Aborted";
     }
     return builds;
@@ -228,7 +228,7 @@ function createCLJSReplType(cljsType: CljsTypeConfig, cljsTypeName: string, conn
     const printThisPrinter: processOutputFn = x => {
             if (cljsType.printThisLineRegExp) {
                 if (x.search(cljsType.printThisLineRegExp) >= 0) {
-                    resultsOutput.appendToResultsDoc('; ' + x.replace(/\s*$/, ""));
+                    outputWindow.appendToResultsDoc('; ' + x.replace(/\s*$/, ""));
                 }
             }
         },
@@ -248,32 +248,32 @@ function createCLJSReplType(cljsType: CljsTypeConfig, cljsTypeName: string, conn
             // When the app is ready to start, say so.
             if (!haveShownStartMessage && cljsType.isReadyToStartRegExp) {
                 if (x.search(cljsType.isReadyToStartRegExp) >= 0) {
-                    resultsOutput.appendToResultsDoc("; CLJS REPL ready to connect. Please, start your ClojureScript app.");
+                    outputWindow.appendToResultsDoc("; CLJS REPL ready to connect. Please, start your ClojureScript app.");
                     haveShownStartMessage = true;
                 }
             }
             // If we have an appURL to go with the ”start now” message, say so
             if (appURL && haveShownStartMessage && !haveShownAppURL) {
                 if (cljsType.shouldOpenUrl) {
-                    resultsOutput.appendToResultsDoc(`; Opening ClojureScript app in the browser at: ${appURL} ...`);
+                    outputWindow.appendToResultsDoc(`; Opening ClojureScript app in the browser at: ${appURL} ...`);
                     open(appURL).catch(reason => {
-                        resultsOutput.appendToResultsDoc("; Error opening ClojureScript app in the browser: " + reason);
+                        outputWindow.appendToResultsDoc("; Error opening ClojureScript app in the browser: " + reason);
                     });
                 } else {
-                    resultsOutput.appendToResultsDoc(";   Open the app on this URL: " + appURL);
+                    outputWindow.appendToResultsDoc(";   Open the app on this URL: " + appURL);
                 }
                 haveShownAppURL = true;
             }
             // Wait for any appURL to be printed before we round of the ”start now” message.
             // (If we do not have the regexp for extracting the appURL, do not wait for appURL.)
             if (!haveShownStartSuffix && (haveShownAppURL || (haveShownStartMessage && !cljsType.openUrlRegExp))) {
-                resultsOutput.appendToResultsDoc(";   The CLJS REPL will connect when your app is running.");
+                outputWindow.appendToResultsDoc(";   The CLJS REPL will connect when your app is running.");
                 haveShownStartSuffix = true;
             }
         },
         // This processor prints everything. We use it for stderr below.
         allPrinter: processOutputFn = x => {
-            resultsOutput.appendToResultsDoc('; ' + util.stripAnsi(x).replace(/\s*$/, ""));
+            outputWindow.appendToResultsDoc('; ' + util.stripAnsi(x).replace(/\s*$/, ""));
         }
 
     let replType: ReplType = {
@@ -342,7 +342,7 @@ function createCLJSReplType(cljsType: CljsTypeConfig, cljsTypeName: string, conn
                         });
                     }
                     if (builds) {
-                        resultsOutput.appendToResultsDoc("; Starting cljs repl for: " + projectTypeName + "...");
+                        outputWindow.appendToResultsDoc("; Starting cljs repl for: " + projectTypeName + "...");
                         state.extensionContext.workspaceState.update('cljsReplTypeHasBuilds', true);
                         startCode = startCode.replace("%BUILDS%", builds.map(x => { return `"${x}"` }).join(" "));
                         const result = evalConnectCode(session, startCode, name, checkFn, [startAppNowProcessor, printThisPrinter], [allPrinter]);
@@ -351,11 +351,11 @@ function createCLJSReplType(cljsType: CljsTypeConfig, cljsTypeName: string, conn
                         }
                         return result;
                     } else {
-                        resultsOutput.appendToResultsDoc("; Aborted starting cljs repl.");
+                        outputWindow.appendToResultsDoc("; Aborted starting cljs repl.");
                         throw "Aborted";
                     }
                 } else {
-                    resultsOutput.appendToResultsDoc("; Starting cljs repl for: " + projectTypeName + "...");
+                    outputWindow.appendToResultsDoc("; Starting cljs repl for: " + projectTypeName + "...");
                     return evalConnectCode(session, startCode, name, checkFn, [startAppNowProcessor, printThisPrinter], [allPrinter]);
                 }
             } else {
@@ -383,20 +383,20 @@ function createCLJSReplType(cljsType: CljsTypeConfig, cljsTypeName: string, conn
 }
 
 async function makeCljsSessionClone(session, repl: ReplType, projectTypeName: string) {
-    resultsOutput.appendToResultsDoc("; Creating cljs repl session...");
+    outputWindow.appendToResultsDoc("; Creating cljs repl session...");
     let newCljsSession = await session.clone();
     newCljsSession.replType = 'cljs';
     if (newCljsSession) {
-        resultsOutput.appendToResultsDoc("; Connecting cljs repl: " + projectTypeName + "...");
-        resultsOutput.appendToResultsDoc(";   The Calva Connection Log might have more connection progress information.");
+        outputWindow.appendToResultsDoc("; Connecting cljs repl: " + projectTypeName + "...");
+        outputWindow.appendToResultsDoc(";   The Calva Connection Log might have more connection progress information.");
         if (repl.start != undefined) {
             if (await repl.start(newCljsSession, repl.name, repl.started)) {
                 state.analytics().logEvent("REPL", "StartedCLJS", repl.name).send();
-                resultsOutput.appendToResultsDoc("; Cljs builds started");
+                outputWindow.appendToResultsDoc("; Cljs builds started");
                 newCljsSession = await session.clone();
             } else {
                 state.analytics().logEvent("REPL", "FailedStartingCLJS", repl.name).send();
-                resultsOutput.appendToResultsDoc("; Failed starting cljs repl");
+                outputWindow.appendToResultsDoc("; Failed starting cljs repl");
                 state.cursor.set('cljsBuild', null);
                 return [null, null];
             }
@@ -409,7 +409,7 @@ async function makeCljsSessionClone(session, repl: ReplType, projectTypeName: st
             let build = state.deref().get('cljsBuild')
             state.analytics().logEvent("REPL", "FailedConnectingCLJS", repl.name).send();
             let failed = "Failed starting cljs repl" + (build != null ? ` for build: ${build}. Is the build running and connected?\n   See the Output channel "Calva Connection Log" for any hints on what went wrong.` : "");
-            resultsOutput.appendToResultsDoc(`; ${failed}`);
+            outputWindow.appendToResultsDoc(`; ${failed}`);
             state.cursor.set('cljsBuild', null);
             vscode.window.showInformationMessage(
                 failed,
@@ -443,7 +443,7 @@ async function promptForNreplUrlAndConnect(port, connectSequence: ReplConnectSeq
             state.cursor.set("port", parsedPort);
             await connectToHost(hostname, parsedPort, connectSequence);
         } else {
-            resultsOutput.appendToResultsDoc("; Bad url: " + url);
+            outputWindow.appendToResultsDoc("; Bad url: " + url);
             util.setConnectingState(false);
             status.update();
         }
@@ -458,7 +458,7 @@ export let nClient: NReplClient;
 export let cljSession: NReplSession;
 export let cljsSession: NReplSession;
 
-export async function connect(connectSequence: ReplConnectSequence, isAutoConnect, isJackIn) {
+export async function connect(connectSequence: ReplConnectSequence, isAutoConnect: boolean) {
     const cljsTypeName = projectTypes.getCljsTypeName(connectSequence);
 
     state.analytics().logEvent("REPL", "ConnectInitiated", isAutoConnect ? "auto" : "manual");
@@ -480,7 +480,7 @@ export async function connect(connectSequence: ReplConnectSequence, isAutoConnec
                 await promptForNreplUrlAndConnect(port, connectSequence);
             }
         } else {
-            resultsOutput.appendToResultsDoc('; No nrepl port file found. (Calva does not start the nrepl for you, yet.)');
+            outputWindow.appendToResultsDoc('; No nrepl port file found. (Calva does not start the nrepl for you, yet.)');
             await promptForNreplUrlAndConnect(port, connectSequence);
         }
     } else {
@@ -489,15 +489,18 @@ export async function connect(connectSequence: ReplConnectSequence, isAutoConnec
     return true;
 }
 
-function standaloneConnect(connectSequence: ReplConnectSequence) {
+async function standaloneConnect(connectSequence: ReplConnectSequence) {
+    await outputWindow.initResultsDoc();
+    const outputDocument = await outputWindow.openResultsDoc();
+
     if (connectSequence) {
         const cljsTypeName = projectTypes.getCljsTypeName(connectSequence);
-        resultsOutput.appendToResultsDoc(`; Connecting ...`);
+        outputWindow.appendToResultsDoc(`; Connecting ...`);
         state.analytics().logEvent("REPL", "StandaloneConnect", `${connectSequence.name} + ${cljsTypeName}`).send();
-        connect(connectSequence, false, false).catch(() => { });
+        connect(connectSequence, false).catch(() => { });
     }
     else {
-        resultsOutput.appendToResultsDoc("; Aborting connect, error determining connect sequence.");
+        outputWindow.appendToResultsDoc("; Aborting connect, error determining connect sequence.");
     }
 }
 
@@ -551,8 +554,8 @@ export default {
                 newSession = util.getSession('cljs');
             }
             state.cursor.set('cljc', newSession);
-            if (resultsOutput.isResultsDoc(vscode.window.activeTextEditor.document)) {
-                resultsOutput.setSession(newSession, undefined);
+            if (outputWindow.isResultsDoc(vscode.window.activeTextEditor.document)) {
+                outputWindow.setSession(newSession, undefined);
                 util.updateREPLSessionType();
             }
             status.update();