+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Changes to Calva.
 
 ## [Unreleased]
+- Fix: [Don't open output window until connect starts](https://github.com/BetterThanTomorrow/calva/issues/707)
 
 ## [2.0.111] - 2020-07-29
 - [Handling ansi code by stripping it](https://github.com/BetterThanTomorrow/calva/issues/696)